Move MSSearchWebpart to Protein module
