+++ protein/src/org/labkey/protein/ProteinModule.java
@@ -26,13 +26,15 @@ import org.labkey.api.files.TableUpdaterFileListener;
 import org.labkey.api.module.DefaultModule;
 import org.labkey.api.module.ModuleContext;
 import org.labkey.api.pipeline.PipelineService;
-import org.labkey.api.protein.annotation.CustomAnnotationSetManager;
-import org.labkey.api.protein.annotation.ProteinAnnotationPipelineProvider;
 import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.protein.ProteinService;
+import org.labkey.api.protein.ProteomicsWebPartFactory;
+import org.labkey.api.protein.annotation.CustomAnnotationSetManager;
+import org.labkey.api.protein.annotation.ProteinAnnotationPipelineProvider;
 import org.labkey.api.protein.fasta.FastaDbLoader;
 import org.labkey.api.protein.query.CustomAnnotationSchema;
 import org.labkey.api.protein.query.ProteinUserSchema;
+import org.labkey.api.protein.search.MSSearchWebpart;
 import org.labkey.api.usageMetrics.UsageMetricsService;
 import org.labkey.api.view.BaseWebPartFactory;
 import org.labkey.api.view.Portal;
@@ -83,6 +85,14 @@ public class ProteinModule extends DefaultModule
                     result.setTitle(CustomProteinListView.NAME);
                     result.setTitleHref(ProteinController.getBeginURL(portalCtx.getContainer()));
                     return result;
+                }
+            },
+            new ProteomicsWebPartFactory(MSSearchWebpart.NAME)
+            {
+                @Override
+                public WebPartView<?> getWebPartView(@NotNull ViewContext portalCtx, @NotNull Portal.WebPart webPart)
+                {
+                    return new MSSearchWebpart();
                 }
             }
         );
